Update model templates and refactoring and first complete genetic algorithm
Added the tensor model templates 'Conv2d' and modified 'Linear' and  'ReLU' to correspond to fit the new props convention. Completed the basis for the the 'modelproperties' and 'modelskeleton' classes. Updated the 'test_graph' file to adapt to the new conventions of 'modelskeleton'.

Added scripts:
-'modelloader' with a class capable of dynamically loading models from a main folder
-'structuremodifiers' containing utilities to simplify and automate the modification of a ModelSkeleton
-'parametermodifers' To change a parameter (not in use)
-'geneticgorithm' containing a simple geneticalgorithm class

The genetic algorithm is working. The model modification and constraints solvers need work